Move the add_documents api model to the right place and expose `batchVectorisationMode` parameter
